+++ CHANGELOG.md
@@ -15,11 +15,19 @@ You can find its changes [documented below](#010-2024-11-20).
 
 This release has an [MSRV][] of 1.82.
 
+### Added
+
+* Add `BLACK`, `WHITE`, and `TRANSPARENT` constants to the color types. ([#64][] by [@waywardmonkeys][])
+
 ## [0.1.0][] (2024-11-20)
 
 This release has an [MSRV][] of 1.82.
 
 This is the initial release.
+
+[@waywardmonkeys]: https://github.com/waywardmonkeys
+
+[#64]: https://github.com/linebender/color/pull/64
 
 [Unreleased]: https://github.com/linebender/color/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/linebender/color/releases/tag/v0.1.0